backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
Pillow==11.2.1
django-ckeditor==6.7.3
django-admin-interface==0.30.1
django-colorfield==0.14.0
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.9.0
django-filter==25.1
django-extensions==4.1
openai==1.58.1
